+++ src/components/bluepad32/include/uni_hid_parser_ds3.h
@@ -19,6 +19,7 @@ limitations under the License.
 #ifndef UNI_HID_PARSER_DS3_H
 #define UNI_HID_PARSER_DS3_H
 
+#include <stdbool.h>
 #include <stdint.h>
 
 #include "uni_hid_parser.h"
@@ -29,5 +30,6 @@ void uni_hid_parser_ds3_init_report(struct uni_hid_device_s* d);
 void uni_hid_parser_ds3_parse_input_report(struct uni_hid_device_s* d, const uint8_t* report, uint16_t len);
 void uni_hid_parser_ds3_set_player_leds(struct uni_hid_device_s* d, uint8_t leds);
 void uni_hid_parser_ds3_set_rumble(struct uni_hid_device_s* d, uint8_t value, uint8_t duration);
+bool uni_hid_parser_ds3_does_name_match(struct uni_hid_device_s* d, const char* name);
 
 #endif  // UNI_HID_PARSER_DS3_H