+++ src/main/main.c
@@ -18,9 +18,11 @@ limitations under the License.
 
 #include <stddef.h>
 
-#include "btstack_port_esp32.h"
-#include "btstack_run_loop.h"
-#include "hci_dump.h"
+#include <btstack_port_esp32.h>
+#include <btstack_run_loop.h>
+#include <hci_dump.h>
+#include <hci_dump_embedded_stdout.h>
+
 #include "sdkconfig.h"
 #include "uni_esp32.h"
 #include "uni_main.h"
@@ -40,6 +42,8 @@ int app_main(void) {
 
     // Configure BTstack for ESP32 VHCI Controller
     btstack_init();
+
+    // hci_dump_init(hci_dump_embedded_stdout_get_instance());
 
     // Init Bluepad32, and does not return.
     uni_main(0, NULL);